requirements.txt
Flask
psycopg2-binary==2.9.10